Add files via upload
New York Times is one of the preeminent news publications around the world. Based out of New York city, it has massive influence in the United States and is third in circulation around the country. With the newspaper being of such importance, we wanted to further understand how the publication affects and shapes the media outlook in the country and how it affects its readers as well. To deep dive into that, we looked into a dataset that had metadata from articles over January through April in 2017 & 2018, and then metadata on the comments posted on articles in those months as well. By sifting through the data, we hoped to get a better understanding of what the paper focuses on and how do its readers engage with their articles.
We first ran analysis on the articles themselves to see over the eight months of data we had, what was the New York Times writing about. Our key analysis was looking into, what were the popular new desk article types that were being published, who the most published authors, and what topics were being discussed the most in those articles. Much in line with what we were expecting, we noticed that Politics was by far the most important topic in the New York Times articles. And most of them seemed to revolve our President Donald J Trump, whether it was articles about his presidency, the 2016 elections and controversies. However interestingly enough, Politics was not the most popular section in the papers. Which showed us that discussions of Donald Trump have surpassed just "political" news, however it has almost become entertainment in itself. Donald Trump will be a topic of discussion, in political news, in business news, in international news, and even in culture. This definitely showed how over his presidency, Donald Trump has waded into conversations on almost every scale, including sports, movies, etc.
Along with the articles, we then ran analysis on the comment sections for the articles as well to get a better sense on who the readers were and how were they engaging with the news stories. We looked into which new desk section were commenters the most active on, which type of articles prompted longer comments, and where the commenters were generally from.
From our analysis we noted that OpEd's comment sections were by far the busiest amongst New York Times article. Opinion pieces are generally written to elicit discussion as it showcases a writer’s own point of view and so this matched our expectations. Some interesting ones of note were that foreign was fairly high up on the list, which shows New York Times has a very diverse and international reader base who keep up with the news around the world. And finally, surprisingly we noticed sports was not as high as we had predicted. However, this data would help suggest that news publications are not where fans come for sports news anymore, as the popularity of sports specific media such as ESPN has risen.
Aside from what the readers were interested in, we wanted to see where majority of the commenters from so we could get a better sense of the reach of New York Times and who their consumer base is. From our analysis, it seems fairly obvious that the New York Times is much popular in the liberal areas of the country such as the Northeast and West Coast. New York City had by far the largest base of commenters, followed by California. The rest of the top list was mostly rounded out by major cities along the coasts. New York Times seems to not really capture an audience in the heartland of the country, which is generally more conservative.
Our research into the New York Times uncovered some interesting information about how the newspaper operates and how its audience engages with their material. The code and results showcased below will help provide a well-rounded view on all the analysis we performed on the subject.


Project Background
We have considered the New York Times comments data set for understanding how the newspaper affects and shapes the media outlook in the United States and how the readers react and engage on it's platform.
Source of the Dataset : Kaggle
Link : https://www.kaggle.com/aashita/nyt-comments (https://www.kaggle.com/aashita/nyt-comments)
The data set consists of two sets of CSV files, Articles and Comments. The metadata for Articles and Comments are present for the following years and months :
Year 2017 : January, Febuary, March, April, May
Year 2018 : January, Febuary, March, April
Articles and Comments data are present in total for 9 months.